Activity now only exists in landscape mode, because portrait only caused problems anyways.
Removed commente out Seebars.

Implemented some temporary Robot commands and big parts of the workLoop

Renamed HSVImage to CleenrImage, which makes a lot more sense actually. Plus I moved a couple of methods in there, to have them where they belog.

Added Log output to the provisorical ContrtolUnit class.

Added a couple of more useful default values to the DetectionParameters.

Added two not yet implemented methods to the FocusedObject-class, which will be used to check how far the object is probably away.

Changed some CVCamerImages to direct RGBA-Mats deeper in the call hierarchy

Added a FrameSizeChangedException. This is not the nicest possible method of doing this. Probably some kind of notification might be a nicer way to do that.